Change Sender range from float to uint8_t and remove Constructor::isReady

+++ source/EngineGpuKernels/GenomeTO.cuh
@@ -274,7 +274,7 @@ struct MemoryGenomeTO
 
 struct SenderGenomeTO
 {
-    float range;
+    uint8_t range;
     int maxTimesSent;
 };
